# docs/spec/api/library/library.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/library/library.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Library"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/library/book": {
      "post": {
        "operationId": "Library_AddBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/libraryAddBookResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/libraryAddBookRequest"
            }
          }
        ],
        "tags": [
          "Library"
        ]
      }
    },
    "/v1/library/book_info/{id}": {
      "get": {
        "operationId": "Library_GetBookInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/libraryGetBookInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          }
        ],
        "tags": [
          "Library"
        ]
      }
    }
  },
  "definitions": {
    "libraryAddBookRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "authorId": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "libraryAddBookResponse": {
      "type": "object",
      "properties": {
        "book": {
          "$ref": "#/definitions/libraryBook"
        }
      }
    },
    "libraryBook": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "authorId": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "libraryGetBookInfoResponse": {
      "type": "object",
      "properties": {
        "book": {
          "$ref": "#/definitions/libraryBook"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}